services/scraper_app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
fastapi==0.75.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
multidict==6.0.2
nest-asyncio==1.5.4
packaging==21.3
prompt-toolkit==3.0.28
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
yarl==1.7.2
